refactor: move awareness check to dedicated module
