+++ tests/performance/load_test.py
@@ -9,7 +9,7 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..",
 
 
 from app import create_app
-from application.services.survey_vector_generator import create_random_vector
+from application.services.awareness_check import create_random_vector
 from database.queries import get_subjects
 
 logger = logging.getLogger(__name__)
